Move shared functions for Portfolio/Aggregating to shared PortalDataService [skip-install]
