+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioOverview/ColumnFormPanel/index.tsx
@@ -39,6 +39,7 @@ export const ColumnFormPanel: FC = () => {
     findMatchingSearchProperty,
     columnMessages
   } = useColumnFormPanel()
+
   return (
     <Panel
       isOpen={context.state.columnForm.isOpen}
