+++ SharePointFramework/PortfolioWebParts/src/data/index.ts
@@ -1,7 +1,6 @@
 import { format } from '@fluentui/react/lib/Utilities'
-import { dateAdd, PnPClientStorage, stringIsNullOrEmpty } from '@pnp/core'
+import { dateAdd, PnPClientStorage } from '@pnp/core'
 import {
-  IItemUpdateResult,
   ISearchResult,
   PermissionKind,
   QueryPropertyValueType,
@@ -22,10 +21,8 @@ import {
   ProjectContentColumn,
   ProjectListModel,
   SPContentType,
-  SPDataSourceItem,
   SPFxContext,
   SPProjectColumnItem,
-  SPProjectContentColumnItem,
   SPProjectItem,
   SPTimelineConfigurationItem,
   TimelineConfigurationModel,
@@ -164,7 +161,11 @@ export class DataAdapter implements IPortfolioWebPartsDataAdapter {
         calculatedLevel = strings.DataSourceLevelProject
       }
       level = level ?? calculatedLevel
-      const columns = await this.fetchProjectContentColumns(category, level)
+      const columns = await this.portalDataService.fetchProjectContentColumns(
+        'PROJECT_CONTENT_COLUMNS',
+        category,
+        level
+      )
       const [views, viewsUrls, columnUrls, levels] = await Promise.all([
         this.fetchDataSources(category, level, columns),
         this.portalDataService.getListFormUrls('DATA_SOURCES'),
@@ -762,65 +763,6 @@ export class DataAdapter implements IPortfolioWebPartsDataAdapter {
     }
   }
 
-  public async fetchProjectContentColumns(dataSourceCategory: string, level?: string) {
-    try {
-      if (stringIsNullOrEmpty(dataSourceCategory)) return []
-      const projectContentColumnsList = this.portalDataService.web.lists.getByTitle(
-        strings.ProjectContentColumnsListName
-      )
-      const columnItems = await projectContentColumnsList.items.select(
-        ...Object.keys(new SPProjectContentColumnItem())
-      )()
-      const filteredColumnItems = columnItems.filter(
-        (col) =>
-          col.GtDataSourceCategory === dataSourceCategory ||
-          (!col.GtDataSourceCategory && !col.GtDataSourceLevel) ||
-          (!col.GtDataSourceCategory && _.contains(col.GtDataSourceLevel, level))
-      )
-      return filteredColumnItems.map((item) => new ProjectContentColumn(item))
-    } catch (error) {
-      throw new Error(format(strings.DataSourceCategoryError, dataSourceCategory))
-    }
-  }
-
-  public async updateProjectContentColumn(
-    columnItem: SPProjectContentColumnItem,
-    persistRenderAs = false
-  ): Promise<IItemUpdateResult> {
-    try {
-      const list = this._sp.web.lists.getByTitle(strings.ProjectContentColumnsListName)
-      const properties: SPProjectContentColumnItem = _.pick(
-        columnItem,
-        [
-          'GtColMinWidth',
-          'GtColMaxWidth',
-          persistRenderAs && 'GtFieldDataTypeProperties',
-          persistRenderAs && 'GtFieldDataType'
-        ].filter(Boolean)
-      )
-      return await list.items.getById(columnItem.Id).update(properties)
-    } catch (error) {
-      throw new Error(error)
-    }
-  }
-
-  public async deleteProjectContentColumn(column: Record<string, any>): Promise<any> {
-    try {
-      const list = this._sp.web.lists.getByTitle(strings.ProjectContentColumnsListName)
-      const items = await list.items()
-      const item = items.find((i) => i.GtManagedProperty === column.fieldName)
-
-      if (!item) {
-        throw new Error(format(strings.ProjectContentColumnItemNotFound, column.fieldName))
-      }
-
-      const itemDeleteResult = list.items.getById(item.Id).delete()
-      return itemDeleteResult
-    } catch (error) {
-      throw new Error(error)
-    }
-  }
-
   public async deleteItemFromList(listName: string, itemId: number): Promise<boolean> {
     try {
       const list = this._sp.web.lists.getByTitle(listName)
@@ -845,32 +787,6 @@ export class DataAdapter implements IPortfolioWebPartsDataAdapter {
       return true
     } catch (error) {
       return false
-    }
-  }
-
-  public async updateDataSourceItem(
-    properties: SPDataSourceItem,
-    dataSourceTitle: string,
-    shouldReplace: boolean = false
-  ): Promise<IItemUpdateResult> {
-    try {
-      const list = this._sp.web.lists.getByTitle(strings.DataSourceListName)
-      const [item] = await list.items.filter(`Title eq '${dataSourceTitle}'`)()
-      if (!item) {
-        throw new Error(format(strings.DataSourceItemNotFound, dataSourceTitle))
-      }
-      if (item.GtProjectContentColumnsId && !shouldReplace) {
-        properties.GtProjectContentColumnsId = [
-          ...item.GtProjectContentColumnsId,
-          properties.GtProjectContentColumnsId
-        ]
-        return await list.items.getById(item.Id).update(properties)
-      } else {
-        properties.GtProjectContentColumnsId = properties.GtProjectContentColumnsId as number[]
-        return await list.items.getById(item.Id).update(properties)
-      }
-    } catch (error) {
-      throw new Error(error)
     }
   }
 }
